cpp: tiff: Correct a number of compiler warnings
- explicitly initialise all pointers
- explictily cast where numeric types differ
- use 64-bit offset type when using bigtiff

+++ cpp/lib/ome/bioformats/tiff/Types.h
@@ -51,7 +51,7 @@ namespace ome
       typedef uint16_t directory_index_type;
 
       /// IFD offset.
-      typedef uint32_t offset_type;
+      typedef uint64_t offset_type;
 
       /// Tag number.
       typedef unsigned int tag_type;
